Module Lib.BlockRuleInlineHTML

Inline HTML

Currently, it supports div and script only, but we are thinking of generalizing it to support any tags, since anyhow the inline html rule should be on only when unsafe mode.

module Div : BlockRule.S

UNSAFE Div inline HTML

module Script : BlockRule.S

UNSAFE Script inline HTML