Module Lib.SpanRuleEscape

Escaped character: HTML special characters, e.g. &, <, etc., are translated to &amp;, &lt;, etc. The following characters should be escaped by backslash in markdown.

[ ] \ ` * # _ { } ( ) + - . !

include SpanRule.S

val is_safe : bool

val first_char : FirstChar.t

val construct : (string -> Typ.span list) -> SpanRule.state -> Typ.span option

Try to apply a rule. The first parameter is the (recursive) function that parses a string as a span list.